requirements.txt
requests
dotenv
streamlit
rapidfuzz
networkx
matplotlib
bs4
numpy
faiss-cpu
pytest
scipy